requirements.txt
alembic
Babel
blinker
Bootstrap-Flask
dnspython
Flask
greenlet
gunicorn
idna
importlib-metadata
itsdangerous
Jinja2
Mako
python-dotenv
requests
SQLAlchemy
zipp
Flask-Migrate
Flask-SQLAlchemy